Tiles to divide the QR into smaller regions
