+++ src/lib.rs
@@ -134,8 +134,9 @@
 
 #![allow(
     clippy::items_after_test_module,
-    unused_variables,
-    dead_code,
+    // FIXME: Uncomment below
+    // unused_variables,
+    // dead_code,
     mixed_script_confusables,
     clippy::suspicious_arithmetic_impl,
     clippy::suspicious_op_assign_impl
