+++ src/reader/mod.rs
@@ -1,5 +1,7 @@
 pub mod binarize;
 mod finder;
+mod fitness;
+mod locate;
 pub mod symbol;
 mod utils;
 
@@ -9,7 +11,8 @@ use finder::{group_finders, locate_finders, FinderGroup};
 
 use binarize::BinaryImage;
 use image::DynamicImage;
-use symbol::{Symbol, SymbolLocation};
+use locate::SymbolLocation;
+use symbol::Symbol;
 
 // Decode result
 //------------------------------------------------------------------------------
@@ -150,9 +153,9 @@ mod reader_tests {
         #[allow(unused_imports)]
         use std::sync::Arc;
 
-        let img_path = std::path::Path::new("./assets/hv1.jpg");
+        let img_path = std::path::Path::new("./assets/lot1.jpg");
 
-        let mut img = image::open(img_path).unwrap().to_luma8();
+        let img = image::open(img_path).unwrap().to_luma8();
 
         let prep_path = std::path::Path::new("assets/prep.png");
         let mut bin_img = BinaryImage::prepare(&img);
@@ -161,22 +164,24 @@ mod reader_tests {
 
         let finders = locate_finders(&mut bin_img);
         dbg!(finders.len());
-        finders.iter().for_each(|f| f.c.highlight(&mut img, image::Rgb([255, 0, 0])));
+        // finders.iter().for_each(|f| f.c.highlight(&mut img, image::Rgb([255, 0, 0])));
 
         let groups = group_finders(&finders);
         dbg!(groups.len());
-        groups.iter().for_each(|g| g.highlight(&mut img));
+        // groups.iter().for_each(|g| g.highlight(&mut img));
 
         let sym_locs = locate_symbols(&mut bin_img, groups);
         dbg!(sym_locs.len());
+        sym_locs.iter().for_each(|sl| sl.highlight(&mut img));
+        // sym_locs[0].highlight(&mut img);
+
         let bin_img = Arc::new(bin_img);
         let mut symbols: Vec<Symbol> =
             sym_locs.into_iter().map(|sl| Symbol::new(bin_img.clone(), sl)).collect::<_>();
-        symbols.iter().for_each(|s| s.highlight(&mut img));
 
-        symbols.iter_mut().enumerate().for_each(|(i, s)| {
-            let _ = dbg!(s.decode());
-        });
+        // symbols.iter_mut().for_each(|s| {
+        //     let _ = dbg!(s.decode());
+        // });
 
         let out_path = std::path::Path::new("assets/detect.png");
         img.save(out_path).unwrap();
