+++ src/reader/utils/homography.rs
@@ -142,8 +142,7 @@ impl Homography {
         Ok(Point { x, y })
     }
 
-    #[cfg(feature = "benchmark")]
-    pub fn raw_map(&self, x: f64, y: f64) -> QRResult<(f64, f64)> {
+    pub fn exact_map(&self, x: f64, y: f64) -> QRResult<(f64, f64)> {
         let xp = self[0] * x + self[1] * y + self[2];
         let yp = self[3] * x + self[4] * y + self[5];
         let w = self[6] * x + self[7] * y + 1.0;
